feat: index.js 및 클래스 수정
