+++ task/baseball-game/minjae/index.js
@@ -1,7 +1,6 @@
 'use strict';
 
-import  BaseballGame  from './main.js';
+import BaseballGame  from './main/BaseballGame.js';
 
 const baseballgame =  new BaseballGame();
 baseballgame.gameInit();
-
